refactor(desktop): 拆出 models 设置子组件与纯函数
将 connect-transport、model-defaults、capabilities combobox 与 model row 组件迁入 settings/models/。
