+++ apps/desktop/src/components/settings-view.tsx
@@ -135,484 +135,30 @@ import { ScrollArea } from "@/components/ui/scroll-area";
 
 export type { SettingsFormState } from "@/components/settings/types";
 
-const defaultCustomModelCapabilities: DesktopModelCapability[] = ["chat", "image"];
+import {
+  connectTransportOptionSummary,
+  connectTransportOptionsForProvider,
+  defaultConnectTransportKind,
+  providerShowsConnectTransportPicker,
+  resolveConnectTransportKindForProvider,
+  resolveCustomConnectApiBase,
+} from "@/components/settings/models/connect-transport";
+import { ModelCapabilitiesCombobox } from "@/components/settings/models/model-capabilities-combobox";
+import {
+  type ModelDefaultAssignments,
+  type ModelDefaultRole,
+  type SettingsModelProfile,
+  canAssignAsActiveModel,
+  canAssignAsImageGenerationModel,
+  canAssignAsLightweightChatModel,
+  canAssignAsVideoGenerationModel,
+  defaultCustomModelCapabilities,
+  getSupportedModelDefaultRoles,
+  modelDefaultActionLabel,
+  normalizeModelCapabilitySelection,
+} from "@/components/settings/models/model-defaults";
+import { ModelSettingsRowButton, ModelSettingsRowWithHover } from "@/components/settings/models/model-settings-row";
 
-type SettingsModelProfile = DesktopSnapshot["config"]["models"][number];
-
-type ModelDefaultAssignments = {
-  activeModel: boolean;
-  imageGenerationModel: boolean;
-  videoGenerationModel: boolean;
-  lightweightChatModel: boolean;
-};
-
-type ModelDefaultRole = keyof ModelDefaultAssignments;
-
-const modelCapabilityOptions: Array<{
-  value: DesktopModelCapability;
-  label: string;
-  labelKey: string;
-  summaryKey: string;
-}> = [
-  { value: "chat", label: "Chat", labelKey: 'settings.capabilityChatLabel', summaryKey: 'settings.capabilityChat' },
-  { value: "image", label: "Image", labelKey: 'settings.capabilityImageLabel', summaryKey: 'settings.capabilityImage' },
-  { value: "video", label: "Video", labelKey: 'settings.capabilityVideoLabel', summaryKey: 'settings.capabilityVideo' },
-  { value: "imageGeneration", label: "Image generation", labelKey: 'settings.capabilityImageGenerationLabel', summaryKey: 'settings.capabilityImageGeneration' },
-  { value: "videoGeneration", label: "Video generation", labelKey: 'settings.capabilityVideoGenerationLabel', summaryKey: 'settings.capabilityVideoGeneration' },
-];
-
-type ConnectTransportOption = {
-  value: DesktopTransportKind;
-  label: string;
-  summaryKey?: string;
-};
-
-const connectTransportOptionCatalog = {
-  chatCompletions: {
-    value: "openai-compatible" as const,
-    label: "Chat Completions API",
-  },
-  messagesApi: {
-    value: "anthropic" as const,
-    label: "Messages API",
-  },
-  bedrockApi: {
-    value: "bedrock" as const,
-    label: "Amazon Bedrock",
-  },
-  responsesApi: {
-    value: "open-responses" as const,
-    label: "Responses API",
-    summaryKey: 'settings.transportResponsesApi',
-  },
-  openResponsesApi: {
-    value: "open-responses" as const,
-    label: "Open Responses API",
-  },
-} satisfies Record<string, ConnectTransportOption>;
-
-function connectTransportOptionsForProvider(provider: DesktopModelProvider): ConnectTransportOption[] {
-  switch (provider) {
-    case "openai":
-    case "xai":
-      return [connectTransportOptionCatalog.chatCompletions, connectTransportOptionCatalog.responsesApi];
-    case "google":
-    case "google-vertex-ai":
-      return [connectTransportOptionCatalog.chatCompletions];
-    case "minimax":
-    case "deepseek":
-    case "xiaomi":
-    case "siliconflow":
-      return [connectTransportOptionCatalog.chatCompletions, connectTransportOptionCatalog.messagesApi];
-    case "alibaba":
-      return [
-        connectTransportOptionCatalog.chatCompletions,
-        connectTransportOptionCatalog.messagesApi,
-        connectTransportOptionCatalog.openResponsesApi,
-      ];
-    case "openrouter":
-    case "custom":
-      return [
-        connectTransportOptionCatalog.chatCompletions,
-        connectTransportOptionCatalog.openResponsesApi,
-        connectTransportOptionCatalog.messagesApi,
-      ];
-    case "volcengine":
-      return [
-        connectTransportOptionCatalog.chatCompletions,
-        { ...connectTransportOptionCatalog.responsesApi, summaryKey: undefined },
-      ];
-    case "amazon-bedrock":
-      return [connectTransportOptionCatalog.bedrockApi];
-    case "azure":
-      return [];
-    default:
-      return [];
-  }
-}
-
-function defaultConnectTransportKind(provider: DesktopModelProvider): DesktopTransportKind {
-  if (provider === "vercel-ai-gateway") {
-    return "open-responses";
-  }
-  if (provider === "amazon-bedrock") {
-    return "bedrock";
-  }
-  if (provider === "azure") {
-    return "open-responses";
-  }
-
-  return connectTransportOptionsForProvider(provider)[0]?.value ?? "openai-compatible";
-}
-
-function providerSupportsConnectTransportPicker(
-  provider: DesktopModelProvider | null,
-): provider is DesktopModelProvider {
-  return (
-    provider === "openai" ||
-    provider === "xai" ||
-    provider === "minimax" ||
-    provider === "deepseek" ||
-    provider === "xiaomi" ||
-    provider === "siliconflow" ||
-    provider === "alibaba" ||
-    provider === "custom" ||
-    provider === "openrouter" ||
-    provider === "volcengine"
-  );
-}
-
-function providerShowsConnectTransportPicker(provider: DesktopModelProvider | null): boolean {
-  return provider !== null
-    && provider !== "vercel-ai-gateway"
-    && providerSupportsConnectTransportPicker(provider);
-}
-
-function resolveConnectTransportKindForProvider(
-  provider: DesktopModelProvider | null,
-  connectTransportKind: DesktopTransportKind,
-): DesktopTransportKind | undefined {
-  if (provider === "vercel-ai-gateway") {
-    return "open-responses";
-  }
-  if (provider === "amazon-bedrock") {
-    return "bedrock";
-  }
-  if (provider === "azure") {
-    return "open-responses";
-  }
-
-  if (provider === null || !providerSupportsConnectTransportPicker(provider)) {
-    return undefined;
-  }
-
-  return connectTransportKind;
-}
-
-function connectTransportOptionSummary(
-  option: ConnectTransportOption,
-  provider: DesktopModelProvider | null,
-): string | undefined {
-  if (option.value === "open-responses" && provider === "xai") {
-    return i18n.t('settings.transportXaiResponses');
-  }
-
-  if (option.value === "open-responses" && provider === "alibaba") {
-    return i18n.t('settings.transportAlibabaResponses');
-  }
-
-  return option.summaryKey ? i18n.t(option.summaryKey) : undefined;
-}
-
-function resolveCustomConnectApiBase(customApiBase: string): string {
-  return customApiBase.trim();
-}
-
-function normalizeModelCapabilitySelection(
-  values: readonly DesktopModelCapability[],
-): DesktopModelCapability[] {
-  const allowed = new Set(modelCapabilityOptions.map((option) => option.value));
-  const seen = new Set<DesktopModelCapability>();
-  const normalized: DesktopModelCapability[] = [];
-  for (const value of values) {
-    if (!allowed.has(value) || seen.has(value)) {
-      continue;
-    }
-    seen.add(value);
-    normalized.push(value);
-  }
-  return normalized.length > 0 ? normalized : [...defaultCustomModelCapabilities];
-}
-
-function canAssignAsActiveModel(
-  model: SettingsModelProfile,
-  isCurrentActiveModel: boolean,
-): boolean {
-  return isCurrentActiveModel || model.capabilities === undefined || model.capabilities.includes("chat");
-}
-
-function canAssignAsImageGenerationModel(
-  model: SettingsModelProfile,
-  isCurrentImageGenerationModel: boolean,
-): boolean {
-  return isCurrentImageGenerationModel || model.capabilities?.includes("imageGeneration") === true;
-}
-
-function canAssignAsVideoGenerationModel(
-  model: SettingsModelProfile,
-  isCurrentVideoGenerationModel: boolean,
-): boolean {
-  return isCurrentVideoGenerationModel || model.capabilities?.includes("videoGeneration") === true;
-}
-
-function canAssignAsLightweightChatModel(
-  model: SettingsModelProfile,
-  isCurrentLightweightChatModel: boolean,
-): boolean {
-  return canAssignAsActiveModel(model, isCurrentLightweightChatModel);
-}
-
-function getSupportedModelDefaultRoles(
-  model: SettingsModelProfile,
-  activeModel: string,
-  imageGenerationModel: string,
-  videoGenerationModel: string,
-  lightweightChatModel: string,
-): ModelDefaultRole[] {
-  const roles: ModelDefaultRole[] = [];
-
-  if (canAssignAsActiveModel(model, model.name === activeModel)) {
-    roles.push("activeModel");
-  }
-
-  if (canAssignAsImageGenerationModel(model, model.name === imageGenerationModel)) {
-    roles.push("imageGenerationModel");
-  }
-
-  if (canAssignAsVideoGenerationModel(model, model.name === videoGenerationModel)) {
-    roles.push("videoGenerationModel");
-  }
-
-  if (canAssignAsLightweightChatModel(model, model.name === lightweightChatModel)) {
-    roles.push("lightweightChatModel");
-  }
-
-  return roles;
-}
-
-function modelDefaultActionLabel(roles: readonly ModelDefaultRole[]): string {
-  if (roles.length === 0) {
-    return i18n.t('settings.noDefaultRoles');
-  }
-
-  if (roles.length === 1) {
-    if (roles[0] === "activeModel") {
-      return i18n.t('settings.setActiveModel');
-    }
-    if (roles[0] === "imageGenerationModel") {
-      return i18n.t('settings.setImageGenModel');
-    }
-    if (roles[0] === "videoGenerationModel") {
-      return i18n.t('settings.setVideoGenModel');
-    }
-    return i18n.t('settings.setLightweightChatModel');
-  }
-
-  return i18n.t('settings.selectDefaultRole');
-}
-
-function ModelCapabilitiesCombobox({
-  value,
-  disabled,
-  onChange,
-}: {
-  value: DesktopModelCapability[];
-  disabled?: boolean;
-  onChange: (value: DesktopModelCapability[]) => void;
-}) {
-  const { t } = useTranslation();
-  const selected = normalizeModelCapabilitySelection(value);
-  const selectedOptions = modelCapabilityOptions.filter((option) =>
-    selected.includes(option.value),
-  );
-  const selectedSet = new Set(selected);
-
-  const toggleCapability = (capability: DesktopModelCapability, checked: boolean) => {
-    const next = checked
-      ? [...selected, capability]
-      : selected.filter((item) => item !== capability);
-    onChange(normalizeModelCapabilitySelection(next));
-  };
-
-  return (
-    <DropdownMenu>
-      <div className={DESKTOP_FORM_INPUT_SHELL}>
-        <DropdownMenuTrigger asChild disabled={disabled}>
-          <button
-            type="button"
-            disabled={disabled}
-            className={cn(
-              DESKTOP_FORM_FIELD_TRIGGER_INNER,
-              "flex min-w-0 items-center justify-between gap-2 py-1 pr-2 pl-1.5 text-sm transition-colors outline-none",
-              "disabled:cursor-not-allowed disabled:opacity-50",
-            )}
-          >
-          <span className="flex min-w-0 flex-1 flex-wrap items-center gap-1">
-            {selectedOptions.length > 0 ? (
-              selectedOptions.map((option) => {
-                const label = t(option.labelKey, { defaultValue: option.label });
-                return (
-                  <span
-                    key={option.value}
-                    className="inline-flex max-w-full items-center gap-1 rounded-md border border-border/60 bg-muted/50 px-1.5 py-0.5 text-xs text-foreground"
-                  >
-                    <span className="truncate">{label}</span>
-                  <span
-                    role="button"
-                    tabIndex={-1}
-                    aria-label={t('settings.removeCapability', { label })}
-                    className="rounded-sm text-muted-foreground hover:text-foreground"
-                    onPointerDown={(event) => {
-                      event.preventDefault();
-                      event.stopPropagation();
-                    }}
-                    onClick={(event) => {
-                      event.preventDefault();
-                      event.stopPropagation();
-                      toggleCapability(option.value, false);
-                    }}
-                  >
-                    <X className="size-3" aria-hidden />
-                  </span>
-                  </span>
-                );
-              })
-            ) : (
-              <span className="px-1 text-muted-foreground">{t('settings.selectCapability')}</span>
-            )}
-          </span>
-          <ChevronsUpDown className="size-4 shrink-0 opacity-60" aria-hidden />
-          </button>
-        </DropdownMenuTrigger>
-      </div>
-      <DropdownMenuContent
-        align="start"
-        sideOffset={6}
-        className="w-[var(--radix-dropdown-menu-trigger-width)]"
-      >
-        {modelCapabilityOptions.map((option) => (
-          <DropdownMenuCheckboxItem
-            key={option.value}
-            checked={selectedSet.has(option.value)}
-            onCheckedChange={(checked) => toggleCapability(option.value, checked)}
-            onSelect={(event) => event.preventDefault()}
-            className="items-start gap-2 py-2"
-          >
-            <span className="min-w-0 flex-1">
-              <span className="block text-sm">{t(option.labelKey, { defaultValue: option.label })}</span>
-              <span className="block truncate text-xs text-muted-foreground">
-                {t(option.summaryKey)}
-              </span>
-            </span>
-          </DropdownMenuCheckboxItem>
-        ))}
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
-
-
-function ModelSettingsRowButton({
-  model,
-  displayTitle,
-  isActive,
-  isImageDefault,
-  isLightweightDefault,
-  defaultActionLabel,
-  disabled,
-  isHighlighted = false,
-  showNativeTitle = true,
-  onPointerEnter,
-  onDefaultAction,
-}: {
-  model: SettingsModelProfile;
-  displayTitle: string;
-  isActive: boolean;
-  isImageDefault: boolean;
-  isLightweightDefault: boolean;
-  defaultActionLabel: string;
-  disabled: boolean;
-  isHighlighted?: boolean;
-  /** 有 HoverDetailTooltip 时不显示浏览器 title，避免盖住详情 Popover。 */
-  showNativeTitle?: boolean;
-  onPointerEnter?: () => void;
-  onDefaultAction: () => void;
-}) {
-  const { t } = useTranslation();
-  const rowAriaLabel = modelSettingsRowAriaLabel(defaultActionLabel, model.name, displayTitle);
-
-  return (
-    <button
-      type="button"
-      className={cn(
-        "flex w-full appearance-none flex-col gap-3 bg-transparent px-4 py-3 text-left outline-none enabled:cursor-pointer enabled:hover:bg-foreground/[0.06] dark:enabled:hover:bg-foreground/10 focus-visible:ring-2 focus-visible:ring-ring/50 sm:flex-row sm:items-center sm:justify-between sm:gap-4",
-        isHighlighted && "bg-muted/30",
-      )}
-      disabled={disabled}
-      title={showNativeTitle ? rowAriaLabel : undefined}
-      aria-label={rowAriaLabel}
-      onPointerEnter={onPointerEnter}
-      onClick={onDefaultAction}
-    >
-      <div className="min-w-0 flex-1">
-        <div className="flex flex-wrap items-center gap-2">
-          <span className="text-sm font-medium text-foreground">{displayTitle}</span>
-          {isActive ? (
-            <Badge variant="secondary" className="text-muted-foreground">
-              {t("settings.currentInference")}
-            </Badge>
-          ) : null}
-          {isImageDefault ? (
-            <Badge variant="secondary" className="text-muted-foreground">
-              {t("settings.currentImageGen")}
-            </Badge>
-          ) : null}
-          {isLightweightDefault ? (
-            <Badge variant="secondary" className="text-muted-foreground">
-              {t("settings.currentLightweightChat")}
-            </Badge>
-          ) : null}
-          {model.capabilities?.map((capability) => (
-            <Badge key={capability} variant="outline" className="text-muted-foreground">
-              {modelCapabilityLabel(capability)}
-            </Badge>
-          ))}
-        </div>
-      </div>
-    </button>
-  );
-}
-
-function ModelSettingsRowWithHover({
-  model,
-  displayTitle,
-  isActive,
-  isImageDefault,
-  isLightweightDefault,
-  defaultActionLabel,
-  disabled,
-  onDefaultAction,
-}: {
-  model: SettingsModelProfile;
-  displayTitle: string;
-  isActive: boolean;
-  isImageDefault: boolean;
-  isLightweightDefault: boolean;
-  defaultActionLabel: string;
-  disabled: boolean;
-  onDefaultAction: () => void;
-}) {
-  const { getTriggerProps } = useHoverDetailTooltipContext<SettingsModelProfile>();
-  const { onPointerEnter, isHighlighted } = getTriggerProps(model);
-
-  return (
-    <HoverDetailTooltip.Anchor itemId={model.name}>
-      <ModelSettingsRowButton
-        model={model}
-        displayTitle={displayTitle}
-        isActive={isActive}
-        isImageDefault={isImageDefault}
-        isLightweightDefault={isLightweightDefault}
-        defaultActionLabel={defaultActionLabel}
-        disabled={disabled}
-        isHighlighted={isHighlighted}
-        showNativeTitle={false}
-        onPointerEnter={onPointerEnter}
-        onDefaultAction={onDefaultAction}
-      />
-    </HoverDetailTooltip.Anchor>
-  );
-}
 
 function ModelsSettingsPanel({
   settings,
